Avoid `RemovedInSphinx80Warning` in path-manipulation code.
In recent versions of Sphinx, code that treats Sphinx application
paths such as `Sphinx.confdir` and `Sphinx.doctreedir` as strings
emits the warning "Sphinx 8 will drop support for representing paths
as strings".

This commit updates path-manipulation code to use functions in the
`os.path` module, which accept path-like objects as well as strings.

This fixes issue #974.